+++ packages/wrapper/web/src/RichContentProps.ts
@@ -1,0 +1,15 @@
+import { EditorState } from 'draft-js';
+export interface RichContentProps {
+  locale?: string;
+  localeResource?: object;
+  placeholder?: string;
+  editorKey?: string;
+  onChange?(editorState: EditorState): void;
+  initialState?: InitialState;
+  theme?: object;
+  config?: object;
+  plugins?: PluginConfig[];
+  ModalsMap?: ModalsMap;
+  helpers?: Helpers;
+  [propName: string]: any;
+}