+++ packages/wrapper/web/src/RichContentTypes.d.ts
@@ -1,0 +1,10 @@
+type InitialState = {
+  blocks: object[];
+  entityMap: { [index: number]: object };
+};
+type Theme = {
+  modalTheme: { content: any };
+  [propName: string]: any;
+};
+type TextToolbarType = 'inline' | 'static';
+type Helpers = { [propName: string]: (...args: any[]) => any };