+++ packages/wrapper/web/src/RichContentWrapper.spec.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 import React from 'react';
-import RichContentWrapper from './RichContentWrapper';
+import { RichContentWrapper } from './RichContentWrapper';
 import { RichContentEditor } from 'wix-rich-content-editor';
 import { RichContentViewer } from 'wix-rich-content-viewer';
 import { pluginHashtag } from '../../../plugin-hashtag/web/src/editor';
@@ -8,7 +8,7 @@ import introState from '../../../../e2e/tests/fixtures/intro.json';
 import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import { default as hebResource } from 'wix-rich-content-common/dist/statics/locale/messages_he.json';
-import { RichContentProps } from './RichContentWrapperTypes';
+import { RichContentProps } from './RichContentProps';
 
 Enzyme.configure({ adapter: new Adapter() });
 const { shallow, mount } = Enzyme;
@@ -32,13 +32,13 @@ const plugins = [pluginHashtag()];
 // eslint-disable-next-line mocha/no-skipped-tests
 describe('Wrapper', () => {
   it('should render editor', () => {
-    const element = shallow(wrapper({ isEditor: true }).withEditor());
+    const element = shallow(wrapper().withEditor());
     expect(element).toBeTruthy();
     expect(element.find('#engine_wrapper')).toBeTruthy();
   });
 
   it('should render editor with locale', () => {
-    const element = shallow(wrapper({ isEditor: true, locale: 'he' }).withEditor());
+    const element = shallow(wrapper({ locale: 'he' }).withEditor());
     expect(element).toBeTruthy();
   });
 
@@ -49,16 +49,16 @@ describe('Wrapper', () => {
 
   describe('Editor', () => {
     it('should render locale="en" if unspecified', () => {
-      const element = mount(wrapper({ isEditor: true }).withEditor());
+      const element = mount(wrapper().withEditor());
       expect(element.props()).toHaveProperty('locale');
       expect(element.props().locale).toEqual('en');
     });
     it('should render editor child if provided', () => {
-      const element = mount(wrapper({ isEditor: true }).withEditor());
+      const element = mount(wrapper().withEditor());
       expect(element.props()).toHaveProperty('children');
     });
     it('should render with pluginsStrategy output', () => {
-      const element = shallow(wrapper({ isEditor: true, plugins }).withEditor());
+      const element = shallow(wrapper({ plugins }).withEditor());
       const instance = element.dive().instance();
       const renderResult = instance.render();
       const editorProps = renderResult.props.children[1].props;
@@ -66,7 +66,7 @@ describe('Wrapper', () => {
       expect(editorProps.config).toHaveProperty('wix-draft-plugin-hashtag');
     });
     it('should render with themeStrategy output', () => {
-      const element = shallow(wrapper({ isEditor: true, theme: 'Default' }).withEditor());
+      const element = shallow(wrapper({ theme: 'Default' }).withEditor());
       const instance = element.dive().instance();
       const renderResult = instance.render();
       const editorProps = renderResult.props.children[1].props;
@@ -74,14 +74,14 @@ describe('Wrapper', () => {
       expect(editorProps.theme).toHaveProperty('modalTheme');
     });
     it('should call updateLocale on componentDidMount', () => {
-      const element = shallow(wrapper({ isEditor: true, locale: 'en' }).withEditor());
+      const element = shallow(wrapper({ locale: 'en' }).withEditor());
       const instance = element.instance();
       const spyUpdate = spyOn(instance, 'updateLocale');
       instance.componentDidMount();
       expect(spyUpdate.calls.count()).toEqual(1);
     });
     it('should render localeStrategy in strategies', async () => {
-      const element = shallow(wrapper({ isEditor: true, locale: 'he' }).withEditor());
+      const element = shallow(wrapper({ locale: 'he' }).withEditor());
       const instance = element.instance();
       const renderResult = instance.render();
       await instance.updateLocale();