+++ packages/wrapper/web/src/themeStrategy/themeStrategy.ts
@@ -1,4 +1,3 @@
-import { FinalTheme } from './../RichContentWrapperTypes';
 import ThemeGenerator from './ThemeGenerator';
 import jss from 'jss';
 import preset from 'jss-preset-default';
@@ -7,7 +6,7 @@ import { defaultTheme } from './defaults';
 export default function themeStrategy(
   isEditor: boolean,
   themeProperties: ThemeProperties
-): FinalTheme {
+): { theme: Theme } {
   const { theme } = themeProperties;
   if (typeof theme === 'object') {
     return { theme: { ...defaultTheme, ...theme } };